# food-retail-marketplace/ardo-backend/docs/swagger.yaml
definitions:
  auth.LoginIn:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.RegisterIn:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      preferredLang:
        type: string
    type: object
...
  current_user.UpdateCurrentLanguageIn:
    properties:
      lang:
        type: string
    type: object
  current_user.UpdateCurrentUserCredentialsIn:
    properties:
      firstName:
        type: string
      lastName:
        type: string
    type: object
  current_user.UpdateCurrentUserEmailIn:
    properties:
      email:
        type: string
    type: object
  current_user.UpdateCurrentUserPasswordIn:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
  current_user.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      preferredLang:
        type: string
      updatedAt:
        type: string
      userType:
        type: string
    type: object
...
  response.Error:
    properties:
      messages:
        example:
        - error message
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
      traceId:
        example: string
        type: string
    type: object
  response.Success:
    properties:
      data: {}
      success:
        example: true
        type: boolean
      traceId:
        example: string
        type: string
    type: object
host: api.ardogroup.org
info:
  contact:
    name: API Support
    url: https://nassyrov.t.me
  description: Detailed info about all endpoints
  title: Ardo Backend API
  version: 1.0.0
paths:
  /auth/confirm/{link}:
    get:
      consumes:
      - application/json
      description: This can only be done by the unauthenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Confirm activation link
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: This can only be done by the unauthenticated user.
      parameters:
      - description: User login object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.LoginIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Logout
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: This can only be done by the unauthenticated user.
      parameters:
      - description: New user object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Register
      tags:
      - Auth
...
  /me:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/current_user.User'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get current user
      tags:
      - Current User
  /me/credentials:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Make order object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/current_user.UpdateCurrentUserCredentialsIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update current user credentials
      tags:
      - Current User
  /me/email:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update current user email object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/current_user.UpdateCurrentUserEmailIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update current user email
      tags:
      - Current User
  /me/password:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update current user password object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/current_user.UpdateCurrentUserPasswordIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update current user password
      tags:
      - Current User
  /me/preferred-lang:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update current user preferred language object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/current_user.UpdateCurrentLanguageIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update current user preferred language
      tags:
      - Current User
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"